+++ app/src/main/java/org/sdc/ttu/sdc_test_game_app/Statistics/PersonalStatsPresenter.java
@@ -1,0 +1,8 @@
+package org.sdc.ttu.sdc_test_game_app.Statistics;
+
+import org.sdc.ttu.sdc_test_game_app.Statistics.PersonalStatsInterface;
+
+
+
+public class PersonalStatsPresenter implements PersonalStatsInterface.Presenter {
+}
