--- app/src/main/java/org/sdc/ttu/sdc_test_game_app/presenter/PersonalStatsPresenter.java
@@ -1,8 +1,0 @@
-package org.sdc.ttu.sdc_test_game_app.presenter;
-
-import org.sdc.ttu.sdc_test_game_app.presenter.interfaces.PersonalStatsInterface;
-
-
-
-public class PersonalStatsPresenter implements PersonalStatsInterface.Presenter {
-}
