--- app/src/main/java/org/sdc/ttu/sdc_test_game_app/view/MainActivity.java
@@ -1,17 +1,0 @@
-package org.sdc.ttu.sdc_test_game_app.view;
-
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        // Start the PersonalStatsFragment here,
-        // This is the base Activity that all the fragments will anchor to.
-    }
-}
