+++ app/src/main/res/layout/activity_login.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="org.sdc.ttu.sdc_test_game_app.view.LoginActivity">
+    tools:context="org.sdc.ttu.sdc_test_game_app.Login.LoginActivity">
 
 
 <LinearLayout
